[SPARK-35824][CORE][TESTS] Move to IntKeyType
